refactor/#15 : EpisodeModel 을 core:model 모듈로 이동
