+++ domain/episode/build.gradle.kts
@@ -1,3 +1,7 @@
 plugins {
 	alias(libs.plugins.mapisode.java.library)
 }
+
+dependencies {
+	implementation(projects.core.model)
+}
